requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-filter==22.1
sqlparse==0.4.2
fontawesomefree==6.2.0
